% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerTypes.R
\name{CancerTypes}
\alias{CancerTypes}
\title{Return a character vector of some common cancer types}
\usage{
CancerTypes()
}
\description{
Return a character vector of some common cancer types
}
\examples{
CancerTypes()[1:5]
}
