% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSMIC.v3.2.R
\docType{data}
\name{COSMIC.v3.2}
\alias{COSMIC.v3.2}
\title{Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer, (v3.2 - March 2021)}
\format{
A list with two elements, \code{signature} and \code{etiologies}.
\itemize{
\item \code{signature} is a list with one element:
\itemize{
\item \code{genome}: A list with the elements:
\itemize{
\item \code{SBS96}: Strand-agnostic single-base substitutions in trinucleotide context.
\item \code{SBS192}: Transcriptionally stranded single-base substitutions in trinucleotide context.
\item \code{DBS78}: Strand-agnostic doublet-base substitutions.
\item \code{ID}: Strand-agnostic indels.
}
}
\item \code{etiologies} is a list with elements:
\itemize{
\item \code{SBS96}
\item \code{SBS192}
\item \code{DBS78}
\item \code{ID}
}

Each element in \code{etiologies} is a single column matrix with rownames being the
signature IDs and values being a short character string
description of the proposed etiology. \cr

In general use \code{\link{GetEtiology}}, which handles new
signatures without elements in \code{etiologies}.
}
}
\source{
Files downloaded from
\url{https://cancer.sanger.ac.uk/signatures/downloads/}, 2021 Sep and
saved in \code{data-raw/COSMIC.v3.2/data/}. \cr Populated by
\code{data-raw/COSMIC.v3.2/code/generate-COSMIC.v3.2-genome-sigs.R}.
}
\usage{
COSMIC.v3.2
}
\description{
Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer, (v3.2 - March 2021)
}
\section{Remark}{

The signatures are all from Human GRCh37 reference genome.
}

\section{Note}{

SBS10c, SBS10d, SBS91, SBS92, SBS93, SBS94 (total 6) new SBS signatures were
added in COSMIC v3.2. See the news from COSMIC release for more details
\url{https://cosmic-blog.sanger.ac.uk/cosmic-mutational-signatures-release-v3-2/}
}

\examples{
SBS96.sigs <- COSMIC.v3.2$signature$genome$SBS96
}
\keyword{datasets}
