% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exposure}
\alias{exposure}
\title{PCAWG7 SigProfiler signature assignments (numbers of mutations due to each signature in each tumor).}
\format{
A list with the elements:

\describe{
  \item{PCAWG}{A list with the elements:
  \describe{
    \item{SBS96}{Strand-agnostic single-base substitutions
                 in trinucleotide context.}
    \item{DBS78}{Strand-agnostic doublet-base substitutions.}
    \item{ID}{Strand-agnostic indels.}
    These are signature assignments for the PCAWG platinum genomes.
  }}
  \item{TCGA}{A list with the elements:
  \describe{
    \item{SBS96}{As above.}
    \item{ID}{As above.}
    These are signature assignments for the TCGA exomes.
  }}
  \item{other.genome}{A list with the element:
  \describe{
     \item{SBS96}{As above.}
     This contains signature assignments for non-TCGA
     genomes.}}
  \item{other.exome}{A list with the element:
  \describe{
     \item{SBS96}{As above.}
     This contains signature assignments for non-TCGA
     exomes.}}
}
}
\source{
Files of \url{https://www.synapse.org/#!Synapse:syn12009743},
  2019 Oct 09, populated by
 \code{data-raw/sig.profiler.sures/load.package.variable.exposure.R}.
}
\usage{
exposure
}
\description{
PCAWG7 SigProfiler signature assignments (numbers of mutations due to each signature in each tumor).
}
\examples{
SBS96.exposure <- exposure$PCAWG$SBS96
}
\keyword{datasets}
