% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExposureInfo.R
\name{ExposureStats1TumorType}
\alias{ExposureStats1TumorType}
\title{Gather exposure statistics for one tumor type}
\usage{
ExposureStats1TumorType(exposure)
}
\description{
Gather exposure statistics for one tumor type
}
\keyword{internal}
