% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEtiology.R
\name{GetEtiology}
\alias{GetEtiology}
\title{Get the proposed etiology of a signature}
\usage{
GetEtiology(mutation.type, sig.id)
}
\arguments{
\item{mutation.type}{character string, one of SBS96, SBS192, DBS78, ID}

\item{sig.id}{character vector with signature ids, e.g.
\code{c("SBS3", "foo")}.}
}
\value{
A character vector of the same length as \code{sig.id},
each element of which is the etiology of the
corresponding signature, if known, or else the empty string.
}
\description{
Get the proposed etiology of a signature
}
\examples{
GetEtiology("ID", c("ID1", "foo", "ID3"))

}
