% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_aliquot_ID_to_SP_ID.R
\name{map_aliquot_ID_to_SP_ID}
\alias{map_aliquot_ID_to_SP_ID}
\title{Translate aliquot IDs (e.g. e0fccaf5-925a-41f9-b87c-cd5ee4aecb59) to "SP" IDs (e.g. SP1682)}
\usage{
map_aliquot_ID_to_SP_ID(aliquot.ids)
}
\arguments{
\item{aliquot.ids}{Character vector of aliquot IDs.}
}
\value{
Character vector of corresponding "SP" IDs. If a corresponding
aliquot ID cannot be found, then return an empty string.
}
\description{
Translate aliquot IDs (e.g. e0fccaf5-925a-41f9-b87c-cd5ee4aecb59) to "SP" IDs (e.g. SP1682)
}
\examples{
aliquot.ids <- c("e0fccaf5-925a-41f9-b87c-cd5ee4aecb59", "foo")
SP.ids <- map_aliquot_ID_to_SP_ID(aliquot.ids)
}
