% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_SP_ID_to_aliquot_ID.R
\name{map_SP_ID_to_aliquot_ID}
\alias{map_SP_ID_to_aliquot_ID}
\title{Translate "SP" IDs (e.g. SP1682) to aliquot IDs (e.g. e0fccaf5-925a-41f9-b87c-cd5ee4aecb59)}
\usage{
map_SP_ID_to_aliquot_ID(SP.ids)
}
\arguments{
\item{SP.ids}{Character vector of "SP" IDs.}
}
\value{
Character vector of corresponding aliquot IDs. If a corresponding
SP ID cannot be found, return NA with a warning.
}
\description{
Translate "SP" IDs (e.g. SP1682) to aliquot IDs (e.g. e0fccaf5-925a-41f9-b87c-cd5ee4aecb59)
}
\details{
If there are SP IDs that cannot be matched to any aliquot IDs,
  return NA with a warning.
}
\examples{
SP.ids <- c("SP58101", "foo")
aliquot.ids <- map_SP_ID_to_aliquot_ID(SP.ids)
}
