% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCAWG.sample.id}
\alias{PCAWG.sample.id}
\title{Vectors of PCAWG \code{tumor_wgs_icgc_specimin_id}.}
\format{An object of class \code{list} of length 3.}
\source{
\url{https://dcc.icgc.org/api/v1/download?fn=/PCAWG/data_releases/latest/release_may2016.v1.4.with_consensus_calls.tsv},
 2019 Oct 09
}
\usage{
PCAWG.sample.id
}
\description{
Note the the PCAWG7 spectra catalogs have 2 blacklisted
sample ids (they were blacklisted after the analysis was
underway). The blacklisted samples are \code{SP116419}
and \code{SP116883}.
}
\keyword{datasets}
