% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCAWG.sample.sheet}
\alias{PCAWG.sample.sheet}
\title{PCAWG sample sheet which contains various sample information}
\format{
A data table with the following columns:
\itemize{
\item donor_unique_id
\item donor_wgs_exclusion_white_gray
\item submitter_donor_id
\item icgc_donor_id
\item dcc_project_code
\item aliquot_id
\item submitter_specimen_id
\item icgc_sample_id
\item dcc_specimen_type
\item library_strategy
}
}
\source{
\url{https://dcc.icgc.org/api/v1/download?fn=/PCAWG/data_releases/latest/pcawg_sample_sheet.v1.4.2016-09-14.tsv},
2019 Oct 15
}
\usage{
PCAWG.sample.sheet
}
\description{
PCAWG sample sheet which contains various sample information
}
\examples{
aliquot.ids <- PCAWG.sample.sheet$aliquot_id

}
\keyword{datasets}
