% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleIdToCancerType.R
\name{SampleIDToCancerType}
\alias{SampleIDToCancerType}
\title{Split out the cancer type from the sample ID for PCAWG IDs}
\usage{
SampleIDToCancerType(PCAWGID)
}
\arguments{
\item{PCAWGID}{A character vector of PCAWG IDs of the form
<cancer.type>::<sample.id>.}
}
\value{
A character vector parallel to \code{PCAWGID} containing
  only the <cancer.type> strings.
}
\description{
Split out the cancer type from the sample ID for PCAWG IDs
}
