% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBS96_ID_to_SBS192_ID.R
\name{SBS96_ID_to_SBS192_ID}
\alias{SBS96_ID_to_SBS192_ID}
\title{Translate SBS96 signature IDs to SBS192 signature IDs by adding "-E" if necessary.}
\usage{
SBS96_ID_to_SBS192_ID(sig.ids)
}
\arguments{
\item{sig.ids}{Character vector of SBS96 signature IDs.}
}
\value{
Character vector of corresponding SBS192 signature IDs; some
  have "-E" (for exome) post-pended.
}
\description{
Translate SBS96 signature IDs to SBS192 signature IDs by adding "-E" if necessary.
}
