% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgnature}
\alias{sgnature}
\alias{signature}
\title{PCAWG7 SigProfiler whole genome signatures.}
\format{An object of class \code{list} of length 2.}
\usage{
signature
}
\description{
This is list with the following elements and sub-elements:
}
\details{
\describe{
  \item{genome}{\describe{
    \item{SBS96}{Strand-agnostic single-base substitutions in trinucleotide context.}
    \item{SBS192}{Transcriptionally stranded single-base substitutions in trinucleotide context}
    \item{DBS}{Strand-agnostic doublet-base substitutions}
  }}
  \item{exome}{\describe{
    \item{SBS96}{Strand-agnostic single-base substitutions in trinucleotide context.}
 }}
}
}
\keyword{datasets}
