% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\docType{data}
\name{signature}
\alias{signature}
\title{Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer, (v3.1 - June 2020)}
\format{

}
\source{
Files downloaded from
  \url{https://cancer.sanger.ac.uk/cosmic/signatures/index.tt}, 2021 Feb and
  saved in \code{data-raw/COSMIC.v3.1/data/}. \cr Populated by
  \code{data-raw/COSMIC.v3.1/code/generate-COSMIC.v3.1-genome-sigs.R}.
}
\usage{
signature
}
\description{
Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
Cancer, (v3.1 - June 2020)
}
\section{Remark}{

The signatures are all from Human GRCh37 reference genome.
}

\keyword{datasets}
