% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spectra}
\alias{spectra}
\title{PCAWG7 mutational spectra (catalogs)}
\format{
A list with the elements:

\describe{
  \item{SBS96}{Deprecated.}

  \item{DBS78}{Deprecated.}

  \item{PCAWG}{A list with the elements:
  \describe{
    \item{SBS96}{Strand-agnostic single-base substitutions
                 in trinucleotide context.}
    \item{SBS192}{Single-base substitutions in transcripts
                 based on the sense strand.}
    \item{SBS1536}{Strand-agnostic single-base substitutions
                 in pentanucleotide context.}

    \item{DBS78}{Strand-agnostic doublet-base substitutions.}
    \item{ID}{Strand-agnostic indels.}

  }}
  \item{TCGA}{A list with the same elements as the
              \code{PCAWG} element.}
  \item{other.genome}{A list with the same elements as the
              \code{PCAWG} element but with \code{ID} omitted.}
  \item{other.exome}{A list with the same elements as the
              \code{PCAWG} element but with \code{ID} omitted.}
}
}
\source{
Files below \url{https://www.synapse.org/#!Synapse:syn11801889},
  2019 Oct 09. Populated by
  \code{data-raw/spectra/load.package.variable.specra.R}.
}
\usage{
spectra
}
\description{
PCAWG7 mutational spectra (catalogs)
}
\examples{
SBS96.spectra <- spectra$PCAWG$SBS96
}
\keyword{datasets}
