% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitExposureByTumorType.R
\name{SplitExposureByTumorType}
\alias{SplitExposureByTumorType}
\title{Split an exposure matrix into a list of matrices, each for a single tumor type.}
\usage{
SplitExposureByTumorType(exposure, tumor.type)
}
\arguments{
\item{exposure}{A numerical matrix or data frame in which rows are
mutational signature and columns are samples (e.g. tumors), and
each element is the number of mutations due to a given mutational
signature in a single sample.}

\item{tumor.type}{A charactor or numeric vector, each element
of which indicates a particular tumor type.}
}
\value{
The list of exposure matrices created by splitting
  \code{exposure} by \code{tumor.type}.
}
\description{
Split an exposure matrix into a list of matrices, each for a single tumor type.
}
