% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitMatrixBySampleType.R
\name{SplitMatrixBySampleType}
\alias{SplitMatrixBySampleType}
\title{Split an exposure matrix or spectrum matrix into a list of matrices, each for a single sample type}
\usage{
SplitMatrixBySampleType(M, sample.type)
}
\arguments{
\item{M}{A numerical matrix or data frame in which
columns are samples (e.g. tumors) and rows are either
mutational signatures (for exposures) or mutation types (for
spectra), and, each element is the number of mutations due
to a given mutational
signature or mutation type in a single sample}

\item{sample.type}{A character or numeric vector, each element
of which indicates a particular sample type.}
}
\value{
Invisibly, the list of exposure or spectrum
  matrices created by splitting
  \code{M} by \code{sample.type}.
}
\description{
Split an exposure matrix or spectrum matrix into a list of matrices, each for a single sample type
}
\examples{
ff <- matrix(1, nrow=3, ncol = 2)
colnames(ff) <- c("sample1", "sample2")
xx <- SplitMatrixBySampleType(ff, c("sample.type.x", "sample.type.y"))
xx

}
