% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitMatrixBySampleType.R
\name{SplitPCAWGMatrixByTumorType}
\alias{SplitPCAWGMatrixByTumorType}
\title{Extract tumor type from column names and return the input matrix split by tumor type.}
\usage{
SplitPCAWGMatrixByTumorType(M)
}
\arguments{
\item{M}{A numerical matrix or data frame or
\code{\link[ICAMS]{ICAMS}} catalog in which
columns are samples (e.g. tumors) and rows are either
mutational signatures (for exposures) or mutation types (for
spectra), and each element is the number of mutations due
to a given mutational
signature or mutation type in a single sample.
The column names must be of the the form
<cancer.type>::<sample.ID>.}
}
\value{
Invisibly, the list of exposure matrices or
   \code{\link[ICAMS]{ICAMS}} catalogs created by splitting
  \code{matrix} by the tumor type encoded in the column names.
}
\description{
Extract tumor type from column names and return the input matrix split by tumor type.
}
\examples{
mm <- SplitPCAWGMatrixByTumorType(spectra$PCAWG$DBS78)
mm[[3]][1:4, 1:5]

}
