% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exposure.stats}
\alias{exposure.stats}
\title{Exposure statistics from the PCAWG7 paper}
\format{
For initial testing see \code{exposure.stats$PCAWG$SBS96}.
}
\source{
Computed from other package variables uisng \code{GatherPCAWG7ExposureStatsSBS96}.
}
\usage{
exposure.stats
}
\description{
Exposure statistics from the PCAWG7 paper
}
\keyword{datasets}
