% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCAWG.sample.id}
\alias{PCAWG.sample.id}
\title{Vectors of the PCAWG \code{tumor_wgs_icgc_specimen_id}s}
\format{
A list with the elements:

\describe{
  \item{white}{Whitelisted IDs}
  \item{grey}{Greylisted IDs}
  \item{black}{Blacklisted IDs}
}
}
\source{
\url{https://dcc.icgc.org/api/v1/download?fn=/PCAWG/data_releases/latest/release_may2016.v1.4.with_consensus_calls.tsv},
 2019 Oct 09
}
\usage{
PCAWG.sample.id
}
\description{
Note that the PCAWG7 spectra catalogs have 2
sample ids that were blacklisted after the mutational
signature analysis was
underway. The blacklisted samples are \code{SP116419}
and \code{SP116883}, which are in \code{PCAWG.sample.id$black}.
}
\examples{
PCAWG.white.ids <- PCAWG.sample.id$white

}
\keyword{datasets}
