% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCAWG.WGS.DBS}
\alias{PCAWG.WGS.DBS}
\title{Doublet Base Substitution (SBS) spectra (deprecated).}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 78 rows and 2780 columns.
}
\usage{
PCAWG.WGS.DBS
}
\description{
Doublet Base Substitution (SBS) spectra (deprecated).
}
\keyword{datasets}
