% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCAWG.WGS.SBS.96}
\alias{PCAWG.WGS.SBS.96}
\title{Single Base Substitution (SBS) spectra in trinucleotide context (deprecated).}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 96 rows and 2780 columns.
}
\usage{
PCAWG.WGS.SBS.96
}
\description{
Single Base Substitution (SBS) spectra in trinucleotide context (deprecated).
}
\keyword{datasets}
