% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG7.R
\docType{package}
\name{PCAWG7}
\alias{PCAWG7}
\title{PCAWG7: A package of data from COSMIC (the Catalogue Of Somatic
Mutations In Cancer) website
\url{https://cancer.sanger.ac.uk/signatures/} and paper 'Repertoire
of Mutational Signatures in Human Cancer'}
\description{
This is a data package with 3 main package variables:
\code{\link{exposure}}, \code{\link{signature}}, and \code{\link{spectra}}.
}
\details{
There are also PDF plots of the signatures in \code{data-raw/plots/}.

There are also several functions for handling PCAWG identifiers:

* \code{\link{map_SP_ID_to_tumor_type}}

* \code{\link{map_aliquot_ID_to_SP_ID}}

* \code{\link{SampleIDToCancerType}}

* \code{\link{SplitPCAWGMatrixByTumorType}}

* \code{\link{SplitMatrixBySampleType}}

The reference for the data is

Alexandrov, L.B., Kim, J., Haradhvala, N.J. et al.
The repertoire of mutational signatures in human cancer.
Nature 578, 94-101 (2020). \url{doi:10.1038/s41586-020-1943-3}.

COSMIC mutational signatures data were downloaded from
\url{https://cancer.sanger.ac.uk/signatures/downloads/}.
}
