% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgnature}
\alias{sgnature}
\alias{signature}
\title{PCAWG7 SigProfiler reference signatures.}
\format{A list with the following elements and sub-elements:

\describe{
  \item{genome}{\describe{A list with the elements:
    \item{SBS96}{Strand-agnostic single-base substitutions in trinucleotide context.}
    \item{SBS192}{Transcriptionally stranded single-base substitutions in trinucleotide context.}
    \item{DBS78}{Strand-agnostic doublet-base substitutions.}
    \item{ID}{Strand-agnostic indels.}
  }}
  \item{exome}{\describe{A list with the elements:
    \item{SBS96}{Strand-agnostic single-base substitutions in trinucleotide context.}
 }}
}}
\source{
Subdirectories of \url{https://www.synapse.org/#!Synapse:syn12009743},
 2019 Oct 09, populated by
 \code{data-raw/populate.variable.siganture.R}.
}
\usage{
signature
}
\description{
PCAWG7 SigProfiler reference signatures.
}
\keyword{datasets}
