% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitExposureByTumorType.R
\name{SplitPCAWGMatrixByTumorType}
\alias{SplitPCAWGMatrixByTumorType}
\title{Extract tumor type from column names and return the input matrix split by tumor type.}
\usage{
SplitPCAWGMatrixByTumorType(M)
}
\arguments{
\item{M}{A numerical matrix or data frame in which
columns are samples (e.g. tumors) and rows are either
mutational signatures (for exposures) or mutation types (for
spectra), and, each element is the number of mutations due
to a given mutational
signature or mutation type in a single sample.}
}
\value{
Invisibly, the list of exposure matrices created by splitting
  \code{matrix} by the tumor type encoded in the column names.
}
\description{
Extract tumor type from column names and return the input matrix split by tumor type.
}
